Fix string parsing
Closes VRVDR-50441

Parsing of certain strings is not being done correctly by the parser used by 'load' and 'cfgcmds'.  This was detected when extending the existing quotes_pass acceptance tests.  Some of the strings could only be in a config file if typed by hand (shell expansion changes the string before they are parsed) but for those in the auto-generated output of show commands, or in saved config, the parser needed to be fixed to handle these.