auth: Define types for TestAutherRequest
A TestAutherRequest can represent either a command authorization or
accounting request for a unit test. Particularly with the addition of
start accounting support it is useful to know exactly the type of
request.

Here we define TestAutherRequestType with known values for command
authorization and stop accounting. The TestAutherRequest constructors
are updated to allow the required type to be specified, and
TestAutherRequestEquals() now additionally checks the type.