auth: AAA, audit, and test TaskAccounter impls.
Update AuditAccounter, AaaAuther, and TestCommandAccounter to define
NewTaskAccounter() which returns a TaskAccounter implementation for
the respective accounting backend.

The existing implementations of AccountCommand() are updated to
delegate through the new TaskAccounter implementations.

An additional TestAutherRequestType type, to represent a start
accounting request, is defined and testTaskAccounter sets the type
accordingly for start/stop accounting requests. This allows UTs to
verify which accounting operations were performed and their ordering.