Rationalise model_extension.go
The ServiceBus() API in model_extension is not only unused, but is never going to be used as VCI comms are handled a different way. Leaving it in only leads to confusion, so I'm removing it. While the Model extension is not actually used, that is left as it is potentially useful, and all other objects are extensible so this matches that expectation.

For now I've left service_bus_test.go in place, but happy to remove it...

Part of VRVDR-50443.