Create ServiceGetState() method.
This commit creates ServiceGetState(), to go with the existing ServiceGetRunning(), ServiceSetRunning() and ServiceValidation() APIs. Longer term plan is to move these out of modelset_extension as they don't belong on it, but short term, this gets the logic for the 4 operations in the same place. Code has been relocated from session_internal.go in configd. Tests will follow in a subsequent configd commit as it relies on some test infra that isn't that easy to backport without a lot of other changes.

NB: I couldn't preserve the 'internal' location of the relocated rfc7951 files as I ended up with a circular dependency via testutils. I figured it was fine to just put the files into the schema directory instead. rfc7951Merger object is private, but New, Tree and Merge methods need to be public for test code to access as test package uses schema_test package so it can use the testutils package (which imports schema). Otherwise we end up with a circular dependency.

Part of VRVDR-50443.