Rework component manager
To allow the component manager to own the ServiceSet/GetRunning, ValidateRunning and GetState methods, we need a ComponentManager object that complies with the ComponentManager interface. The 'real' component manager used from configd will provide the real ServiceManager and the VCI Client as the OperationsManager, whereas the TestComponentManager will use the core ComponentManager object but with test objects providing the Service and Operations Manager interfaces that allow us to test that calls to these APIs are made correctly by the core ComponentManager object.

The TestComponentManager object is reworked to use this API in this commit. The configd 'real' ComponentManager is reworked in the associated configd PR.

Part of VRVDR-50443.