Move ms.ServiceValidation() to compMgr.ComponentValidation()
First step in moving the 4 'Service' methods out of modelset_extension.go / modelset object into the component manager object. Here ServiceValidation() is moved, and the 2 ListActive[OrConfigured]Models methods are copied and reworked. createComponentMappings() is also added to encapsulate this code and remove it from the modelset_extension.go file.

Part of VRVDR-50443.