Remove componentMappings from modelSet; make into more encapsulated object.
Last stage of relocation of code from modelset to componentManager object. All remaining references to componentMappings on the modelset are removed as these are now fully contained in the componentManager object. Code for managing the component mappings is moved to namespace_map.go, and the modelset_test.go file is renamed as namespace_map_test.go as that is what it tests.

The CompilationExtensions object is now reduced to an empty struct as we no longer need to pass the component mappings through it. I have left the empty struct in for now to avoid further replumbing, and any possibility it may be needed in future, but am happy to pull it out completely if preferred (probably as a separate PR).

Part of VRVDR-50443.