Disable go vet for Go1.10 and earlier.
Following the move of modelset actions to the ComponentManager object, go vet is wrongly flagging modelset as not having the ConfigdExt method. This is in the hasExtensions interface, included in the ExtendedNode interface, and implemented by *extensions object which is most definitely included in the modelset object. go vet only complains in Go1.10 and earlier (Debian9 / xUbuntu18.04) and is ok in later versions. See https://github.com/golang/go/issues/22701.

As production builds all use more recent versions of Go, this only impacts legacy Vyatta:Tools builds that we have to keep running for the few laggards still using older Debian / Ubuntu versions. Simplest is just to selectively disable go vet - as we run it on production code, we aren't losing any checking.